tests: separate test app initialization from TestCase (#23)
* chore: create the trait Orkestra\Testing\Traits\HasApplicationTrait
